Move parent class LcnEntity to entity.py
