+++ homeassistant/components/lcn/__init__.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable
 from functools import partial
 import logging
 
@@ -11,26 +10,19 @@ from pypck.connection import PchkConnectionManager
 
 from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
 from homeassistant.const import (
-    CONF_ADDRESS,
     CONF_DEVICE_ID,
-    CONF_DOMAIN,
     CONF_IP_ADDRESS,
-    CONF_NAME,
     CONF_PASSWORD,
     CONF_PORT,
-    CONF_RESOURCE,
     CONF_USERNAME,
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import device_registry as dr
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.typing import ConfigType
 
 from .const import (
     ADD_ENTITIES_CALLBACKS,
     CONF_DIM_MODE,
-    CONF_DOMAIN_DATA,
     CONF_SK_NUM_TRIES,
     CONNECTION,
     DOMAIN,
@@ -38,12 +30,9 @@ from .const import (
 )
 from .helpers import (
     AddressType,
-    DeviceConnectionType,
     InputType,
     async_update_config_entry,
     generate_unique_id,
-    get_device_connection,
-    get_device_model,
     import_lcn_config,
     register_lcn_address_devices,
     register_lcn_host_device,
@@ -239,74 +228,3 @@ def _async_fire_send_keys_event(
                 event_data.update({CONF_DEVICE_ID: device.id})
 
             hass.bus.async_fire("lcn_send_keys", event_data)
-
-
-class LcnEntity(Entity):
-    """Parent class for all entities associated with the LCN component."""
-
-    _attr_should_poll = False
-    device_connection: DeviceConnectionType
-
-    def __init__(
-        self,
-        config: ConfigType,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the LCN device."""
-        self.config = config
-        self.config_entry = config_entry
-        self.address: AddressType = config[CONF_ADDRESS]
-        self._unregister_for_inputs: Callable | None = None
-        self._name: str = config[CONF_NAME]
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID."""
-        return generate_unique_id(
-            self.config_entry.entry_id, self.address, self.config[CONF_RESOURCE]
-        )
-
-    @property
-    def device_info(self) -> DeviceInfo | None:
-        """Return device specific attributes."""
-        address = f"{'g' if self.address[2] else 'm'}{self.address[0]:03d}{self.address[1]:03d}"
-        model = (
-            "LCN resource"
-            f" ({get_device_model(self.config[CONF_DOMAIN], self.config[CONF_DOMAIN_DATA])})"
-        )
-
-        return DeviceInfo(
-            identifiers={(DOMAIN, self.unique_id)},
-            name=f"{address}.{self.config[CONF_RESOURCE]}",
-            model=model,
-            manufacturer="Issendorff",
-            via_device=(
-                DOMAIN,
-                generate_unique_id(
-                    self.config_entry.entry_id, self.config[CONF_ADDRESS]
-                ),
-            ),
-        )
-
-    async def async_added_to_hass(self) -> None:
-        """Run when entity about to be added to hass."""
-        self.device_connection = get_device_connection(
-            self.hass, self.config[CONF_ADDRESS], self.config_entry
-        )
-        if not self.device_connection.is_group:
-            self._unregister_for_inputs = self.device_connection.register_for_inputs(
-                self.input_received
-            )
-
-    async def async_will_remove_from_hass(self) -> None:
-        """Run when entity will be removed from hass."""
-        if self._unregister_for_inputs is not None:
-            self._unregister_for_inputs()
-
-    @property
-    def name(self) -> str:
-        """Return the name of the device."""
-        return self._name
-
-    def input_received(self, input_obj: InputType) -> None:
-        """Set state/value when LCN input object (command) is received."""
