+++ homeassistant/components/lcn/climate.py
@@ -25,7 +25,6 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType
 
-from . import LcnEntity
 from .const import (
     ADD_ENTITIES_CALLBACKS,
     CONF_DOMAIN_DATA,
@@ -35,6 +34,7 @@ from .const import (
     CONF_SETPOINT,
     DOMAIN,
 )
+from .entity import LcnEntity
 from .helpers import InputType
 
 PARALLEL_UPDATES = 0
