+++ homeassistant/components/lcn/cover.py
@@ -13,7 +13,6 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType
 
-from . import LcnEntity
 from .const import (
     ADD_ENTITIES_CALLBACKS,
     CONF_DOMAIN_DATA,
@@ -21,6 +20,7 @@ from .const import (
     CONF_REVERSE_TIME,
     DOMAIN,
 )
+from .entity import LcnEntity
 from .helpers import InputType
 
 PARALLEL_UPDATES = 0
