+++ homeassistant/components/lcn/sensor.py
@@ -19,7 +19,6 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType
 
-from . import LcnEntity
 from .const import (
     ADD_ENTITIES_CALLBACKS,
     CONF_DOMAIN_DATA,
@@ -30,6 +29,7 @@ from .const import (
     THRESHOLDS,
     VARIABLES,
 )
+from .entity import LcnEntity
 from .helpers import InputType
